Rename `curves` module to `curve`
This better matches the names of its sibling modules.
